Cleanup of PR #555 merge

+++ kraken/lib/pretrain/util.py
@@ -139,7 +139,7 @@ def compute_mask_indices(shape: Tuple[int, int], mask_prob: float, mask_length: 
         for length in sorted(lengths, reverse=True):
             lens = np.fromiter(
                 (e - s if e - s >= length + mask_min_space else 0 for s, e in parts),
-                np.int,
+                int,
             )
             l_sum = np.sum(lens)
             if l_sum == 0:
